requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
google-cloud-dialogflow==2.24.1
google-auth==2.23.3
requests==2.31.0
python-dotenv==1.0.0
